refactor(profiles): Profile is the only real entity again
